refactor: refactor the code for validating permissions when modifying the public status using `@ModificationPermissionCheck` annotation

+++ src/main/java/com/github/learndifferent/mtm/mapper/BookmarkMapper.java
@@ -223,7 +223,7 @@ public interface BookmarkMapper {
      * @param id ID
      * @return bookmark
      */
-    BookmarkDO getBookmarkById(Integer id);
+    BookmarkDO getBookmarkById(long id);
 
     /**
      * Retrieve the bookmark by ID
